remove unused interfaces
