+++ packages/core/src/Helpers/index.ts
@@ -1,3 +1,3 @@
 export { RxBuilder, RxConfig, DestroyAction } from './RxBuilder';
-export { Cases } from './createSlice';
+export { Cases, Reducer } from './createSlice';
 export { combine } from './combine';
