+++ packages/core/src/Models/index.ts
@@ -1,4 +1,3 @@
 export { Action, ScopedEffects } from './Action';
 export { Effect } from './Effect';
-export { Hub, Reducer, HubConfig, StoreConfig, Dispatcher } from './Hub';
 export { Reactable, ActionMap } from './Reactable';
